// src/app/types/asset.ts
export interface AssetType {
    _id?: any;        // اختیاریه چون ممکنه هنگام افزودن هنوز نداشته باشه
    assetType: any;        // اسم دارایی (مثل طلا، دلار، خودرو، بیت‌کوین و...)
    assetAmount: any;      // مقدار دارایی (مثل 12 گرم)
    assetAmountType: any;        // واحد مقدار (گرم، عدد، متر، سهم و ...)
  }
  